計算、描画処理の追加

+++ app/src/main/java/systems/ohno/h_money/model/Speed.kt
@@ -2,5 +2,5 @@ package systems.ohno.h_money.model
 
 
 data class Speed (
-        val hour_speed: Int
+        val hour_speed: Long
 )
